requirements.txt
streamlit==1.37.0
langchain==0.2.14
langchain-community==0.2.12
langchain-groq==0.1.4
chromadb==0.4.24
sentence-transformers==2.5.1
huggingface-hub==0.23.0
pypdf==4.2.0
python-dotenv==1.0.1

